Remove the coprocessor subproject
After kickoff and the FIRST POWER UP game reveal we have decided that an
extra coprocessor like the Raspberry Pi is not necessary because of how
small the importance of vision is in this year's game. Because of this,
we can drop the extra build files and code for a coprocessor. This also
decreases complexity in the project, a good thing for new people.
