+++ build.gradle
@@ -1,28 +1,38 @@
-// Note 'apply false' for each plugin. This ensures they aren't applied to the root project (here), 
-// but it makes the plugins available for subprojects.
 plugins {
-    // GradleRIO is used for code intended to go on robots, which is why it's only applied to the 'robot' project.
-    id "jaci.openrio.gradle.GradleRIO" version "2018.01.06b" apply false
-
-    // EmbeddedTools is the backing library for GradleRIO, and includes everything that's non-frc specific, which
-    // is why we use it for our coprocessor project.
-    id "jaci.gradle.EmbeddedTools" version "2017.12.23" apply false
+    id "idea"
+    id "eclipse"
+    id "jaci.openrio.gradle.GradleRIO" version "2018.01.06b"
+    id "java"
 }
 
-// Using the 'idea' plugin allows us to use IntelliJ IDEA. IntelliJ works out-of-the-box with multiproject builds
-// and so it's very, very easy to integrate with GradleRIO and EmbeddedTools.
-// We need to apply 'idea' to all projects, including the root project (here).
-allprojects {
-    apply plugin: "idea"
-    apply plugin: "eclipse"
+def TEAM = 1294
+def ROBOT_CLASS = 'org.team1294.firstpowerup.robot.Robot'
+
+deploy {
+    targets {
+        target("roborio", jaci.openrio.gradle.frc.RoboRIO) {
+            team = TEAM
+        }
+    }
+    artifacts {
+        artifact('robotCode', jaci.openrio.gradle.frc.FRCJavaArtifact) {
+            targets << "roborio"
+        }
+    }
 }
 
-// Setup the configuration for all sub-projects. This is a shortcut for manually applying the configuration
-// in each subproject individually.
-subprojects {
-    apply plugin: "java"
+dependencies {
+    compile wpilib()
+    compile ctre()
+    compile navx()
+}
+
+jar {
+    from configurations.compile.collect { it.isDirectory() ? it : zipTree(it) }
+    manifest jaci.openrio.gradle.GradleRIOPlugin.javaManifest(ROBOT_CLASS)
 }
 
 task wrapper(type: Wrapper) {
     gradleVersion = '4.2.1'
 }
+
