--- coprocessor/build.gradle
@@ -1,36 +1,0 @@
-apply plugin: 'jaci.gradle.EmbeddedTools'
-
-deploy {
-    targets {
-        target("rpi") {
-            addresses << 'raspberrypi'
-            user = 'pi'
-            password = 'raspberry'
-            directory = '/home/pi/mycode'       // Note: must be an absolute path
-            failOnMissing = false   // Just in case our coprocessor is missing. Optional.
-        }
-    }
-    artifacts {
-        // Define my artifact. In this case, the default java build for this project.
-        javaArtifact('coprocessorCode') {
-            targets << "rpi"
-            jar = 'jar'     // This sets the jarfile to deploy. The default jar task in gradle is simply called 'jar'
-        }
-    }
-}
-
-repositories {
-    mavenCentral()
-}
-
-dependencies {
-    compile 'junit:junit:4.12'
-}
-
-jar {
-    from configurations.compile.collect { it.isDirectory() ? it : zipTree(it) }
-    // I need to set my manifest up so the coprocessor knows where my main class is
-    manifest {
-        attributes 'Main-Class': 'org.team1294.firstpowerup.coprocessor.CoprocessorExample'
-    }
-}
