--- robot/build.gradle
@@ -1,30 +1,0 @@
-apply plugin: "jaci.openrio.gradle.GradleRIO"
-
-deploy {
-    targets {
-        target("roborio", jaci.openrio.gradle.frc.RoboRIO) {
-            team = 1294
-        }
-    }
-    artifacts {
-        // Define my robot artifact. The FRCJavaArtifact type gives access to
-        // robotCommand and other tools necessary for the RoboRIO to detect
-        // and run our code.
-        artifact('robotCode', jaci.openrio.gradle.frc.FRCJavaArtifact) {
-            targets << "roborio"
-        }
-    }
-}
-
-// Defining my dependencies. In this case, WPILib (+ friends), CTRE Toolsuite (Talon SRX)
-// and NavX.
-dependencies {
-    compile wpilib()
-    compile ctre()
-    compile navx()
-}
-
-jar {
-    from configurations.compile.collect { it.isDirectory() ? it : zipTree(it) }
-    manifest jaci.openrio.gradle.GradleRIOPlugin.javaManifest('org.team1294.firstpowerup.robot.Robot')
-}
