+++ src/main/java/org/team1294/firstpowerup/robot/RobotMap.java
@@ -1,0 +1,12 @@
+package org.team1294.firstpowerup.robot;
+
+/**
+ * A map of constants for ports and other magic numbers in the code. Other
+ * classes refer to these constants (i.e. {@code RobotMap.SOME_CONSTANT}.
+ */
+public class RobotMap {
+    public static final int JOYSTICK_DRIVE = 0;
+
+    public static final int EXAMPLE_BUTTON = 0;
+    public static final int EXAMPLE_MOTOR_CONTROLLER = 0;
+}
